+++ ui/src/components/BeamlineActions/BeamlineActionDialog.jsx
@@ -1,22 +1,26 @@
 import React from 'react';
-import { Row, Col, Modal, Button, Form, Card } from 'react-bootstrap';
+import { Row, Col, Modal, Button, Card } from 'react-bootstrap';
 import Plot1D from '../Plot1D';
 import { DraggableModal } from '../DraggableModal';
+import AnnotatedBeamlineActionForm from './AnnotatedBeamlineActionForm';
+import BeamlineActionForm from './BeamlineActionForm';
 
-export default function BeamlineActionDialog(props) {
+export default function AnnotatedBeamlineActionDialog(props) {
   const {
     isDialogVisble,
     handleOnHide,
     defaultPosition,
     actionName,
     actionId,
+    actionType,
     actionArguments,
+    actionSchema,
     isActionRunning,
     actionMessages,
-    handleSetActionArgument,
     handleStopAction,
     handleStartAction,
     handleOnPlotDisplay,
+    handleSetActionArgument,
     plotId,
   } = props;
 
@@ -30,73 +34,42 @@ export default function BeamlineActionDialog(props) {
       <Modal.Header>
         <Modal.Title>{actionName}</Modal.Title>
       </Modal.Header>
-      <Modal.Body
-        className="d-flex"
-        style={{ height: '500px', overflowY: 'auto' }}
-      >
-        <Row>
-          {actionArguments.map((arg, i) => (
-            <Fragment key={arg.name}>
-              <Col
-                className="mt-2"
-                xs={3}
-                style={{ whiteSpace: 'nowrap' }}
-                component={Form.Label}
-              >
-                {arg.name}
-              </Col>
-              <Col xs={3}>
-                <Form.Control
-                  label={arg.name}
-                  type="text"
-                  value={arg.value}
-                  disabled={isActionRunning}
-                  onChange={(e) => {
-                    handleSetActionArgument(actionId, i, e.target.value);
-                  }}
-                />
-              </Col>
-            </Fragment>
-          ))}
+      <Modal.Body>
+        {actionType === 'List' && (
+          <BeamlineActionForm
+            actionId={actionId}
+            isActionRunning={isActionRunning}
+            actionArguments={actionArguments}
+            handleStopAction={handleStopAction}
+            handleStartAction={handleStartAction}
+            handleSetActionArgument={handleSetActionArgument}
+          />
+        )}
+        {actionType === 'JSONSchema' && (
+          <AnnotatedBeamlineActionForm
+            actionId={actionId}
+            actionSchema={actionSchema}
+            isActionRunning={isActionRunning}
+            handleStopAction={handleStopAction}
+            handleStartAction={handleStartAction}
+          />
+        )}
+        <Row className="py-2">
           <Col>
-            {isActionRunning ? (
-              <Button
-                variant="danger"
-                onClick={() => {
-                  handleStopAction(actionId);
-                }}
-              >
-                Abort
-              </Button>
-            ) : (
-              <Button
-                disabled={isActionRunning}
-                variant="primary"
-                onClick={() => {
-                  handleStartAction(actionId);
-                }}
-              >
-                Run
-              </Button>
+            <Plot1D
+              displayedPlotCallback={handleOnPlotDisplay}
+              plotId={plotId}
+              autoNext={isActionRunning}
+            />
+            {actionMessages.length > 0 && (
+              <Card>
+                {actionMessages.map((message) => (
+                  <p key={message}>{message.message}</p>
+                ))}
+              </Card>
             )}
           </Col>
         </Row>
-        <hr />
-        <Plot1D
-          displayedPlotCallback={handleOnPlotDisplay}
-          plotId={plotId}
-          autoNext={isActionRunning}
-        />
-        {actionMessages.length > 0 ? (
-          <Card>
-            {actionMessages.map((message) => (
-              // eslint-disable-next-line react/jsx-key
-              <p>{message.message}</p>
-            ))}
-          </Card>
-        ) : (
-          ''
-        )}
       </Modal.Body>
       <Modal.Footer>
         <Button
